requirements.txt
asgiref==3.6.0
boto3==1.26.133
botocore==1.29.133
Django==4.2
django-storages==1.13.2
environ==1.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
